rename class in speaking scss
