+++ src/pages/Test/Speaking/Speaking.scss
@@ -12,7 +12,7 @@
     cursor: pointer;
   }
 
-  .off {
+  .microphone-off {
     background-color: #d7463e;
   }
